Created single file for the entire free worlds story
This is to make specific missions more easily searchable, set all story files next to each other, and declutter the data folder. FW specific events and jobs are also to be handled here for ease of editing/modification and for player searching.

Basically this just removes the headache of having 7 different places to look.