First draft weapons rebalance
This is far from complete or playable, but this should be fine for testing and refining. 

The biggest changes are how gimmicks are applied, and how Torpedoes work. All human space weapons should now have a gimmick, and torpedoes should only last about 2 seconds and travel even slower, however they will ignore 90% of the targets shield assuming they hit. This will make fighters and gunships better glass cannons, so they are more of a threat in game. 

This correction should also correct several of the "suicidal pirates!" complaints, as now even the smallest ship should be capable of holding its own in a big fleet brawl.