Decreased cust of heavy shuttles
This is a theoretical (untested) fix to the shuttle start. 

The main thing here is to make the shuttle start the most expensive start, but also replace the standard shuttle with the heavy shuttle. This should, in theory, fix the start for people who are playtesting the currently unplayable ESCW