Cloaking update to smaller ships
Smaller ships with cloaking now cloak for between 3-7 seconds while at full forward thrust in their stock configuration. 

The lack of true balance here is currently to allow players more things to play with, however if this creates a new meta or a lopsided balance situation, this will be changed in the future.