Rebalanced weapons
All primary cannons should have an absolute disable of ~30 seconds on a stock behemoth, and now they do. This should allow players to pick what kind of play style they would like to run, rather than having an absolute meta. 

This of course is still rough, and will likely require more testing and rebalancing. But this is a start.