Poisonwood: Alpha Fixes
Unfortunately the politics surrounding the Alphas were less than well executed, which lead to some debate as to how best to correct this. Everyone polled agreed that it needed changed, but it was a three way tie on how to best fix the issue - between adding new dialogue, changing the lore, and just cutting out Poisonwood from the story all together. This edit is the closest thing to an answer that would please everyone as I could get. 

This removes: 
All mention of how Alphas breed (It was literally rape by artificial insemination)
The morality of essentially forcing these women to have an abortion against their will, after they were impregnated against their will. 
That whole political hotbutton mess

This adds:
A different political hotbutton mess - the right to death with dignity 
A more sinister motive behind the capture of the planet than "breeding." 
Three (3) ships disabled in orbit of the planet. Two of which are Korath designs. All three are able to be captured, but the Korath vessels have an 85% chance to self destruct and are not easy to take
A subtle reference to the fact that there are more types of Augments than just Alphas. Not every faction would have bred exactly the same kind of super soldier, after all. (Renamed Alpha Wars to Augment Wars)

This hints at: 
A potential for different alpha/augment factions to be in opposition with each other
An idea one of the Alpha factions might be playing humanity against each other, and against the Alphas you just took
The fact that the Alphas are incredibly powerful, and the only benefits Humanity has over them are numbers and production capability. 

This also preserves the correction to assuming the player isn't Colorblind. 1 in 12 men and 1 in 200 women world wide are colorblind, with the most common deficient color being green. We should expect some of our players to be colorblind, so I gave the alphas a feature that does not assume people could tell if they had green skin.