Story Resources File
Adds new file for "unnamed," "unknown," and "untitled" ships/resources by the Alphas. These will often be the same as those used by other factions in Invasion on, but will not have their proper title. 

This is to foreshadow things without spoiling them, and without creating logical inconsistencies - such as how Human scanners would be able to read the name and designations of Kor Automa ships.