Implemented jafdy Fix and added choice dialogue
jafdy made the very strong argument that nothing that goes into a Neutron Bomb is actually all that dangerous, so the implications of high dose radiation exposure make no sense unless the hostages were being tested on. 

He also mentioned how the Alphas would have use for a very small N-bomb, and at this era it could even be compact enough to be used as a personal weapon. Especially considering the Alpha's new resistance to radiation. 

Added clarity that the resources gathered were not specifically for the N-bombs themselves, but rather for the delivery systems of the N-bombs.