Added Veles to ships
The Veles will be a new Battle Cruiser similar to the Tarazed Vulture. It will not be for sale, nor will it be capturable, and will be in the high tier 2 ranges. 

This ship is intended to play a key part in one of the Northern Union branches, as well as one of the Republic Navy branches. 

Spoilers, it's stupidly powerful. Like, I'm keeping it out of the hands of players for a reason.