Update outfitters, remove more aliens
Ammo should be sold on all planets with outfitters (yes, this includes gatling ammo for a reason)
As ESS will launch with just its base game (ESCW) attached, all alien planets/existing content that we want to keep/revise still should not be included in the base game. Those will instead be released as official plugins. So more and more alien content will be cleaned up as time progresses.