allow me my soapbox
i need to scream about how Github needs a better web interface real quick. The web interface was the source of this problem in the first place (not defaulting to dropping to the folder i intended to create without typing in the address, which is an easy thing to miss) then forcing me to delete each individual file one at a time by hand because theres no quick revert button, or way to check off specific files for mass deletion. 

I shouldn't have to open up gitbash or use a desktop application to do this efficiently if the web interface exists. It exists for a reason. Let me use it please. 

Thank you.