added first part of Marina arc
The NU needs a "katya esque" arc, and nothing says "katya esque" like step for step ripping off the first part of her arc before diving deep into a unique story. This is the step for step ripoff thus far, but there will be more - and that more will be unique.