Remove Nomos and Arete from fleets
MZ stated at one point that there are maybe 80 "carriers" in the whole republic navy period. Reason dictates there are probably not very many cruisers either, however the frigates have been in mass production for a long time, and it makes sense for them to be numerous. This is intended to work with Unique Fleets and the change made to the Frigate to allow it to carry drones.