Moved Thrall, added stock exchange
Thrall essentially just cluttered the pollux system and offered nothing except a "stark contrast" between itself and martini. 
Moved it one system over, so the contrast is preserved without the clutter. 

Added martini stock exchange, this station is intended to be removed on the war begins event, as it will be hit by a station killer and dropped on the planet below.