make bourne useful
come on, its the capital of the free worlds. Give people a reason to actually go there. 

Also, at this point in the story, you are locked out of most content by prohibitively difficult and long journeys to other regions of space. This allows players more flexibility in their fleets, in a lore friendly way.