So, so very much
Sorry @RestingImmortal in advance if this creates a mess of errors.

Oh my this was actually a lot of work, but a lot less than I was expecting it to be:
Cut all alien related content
Fixed the missions so they should work without the alien content
Decreased the difficulty of finding the missions as the Baccy rebalances no longer require this to be hidden behind stacks of obscurity
Decreased the difficulty of combat, as this should no longer be difficult
Referenced the NU where required
Added flavor
Cut pointless content and fluff
Added more flavor
Cut more content

And did all this whist trying to stay as true to Derpy's vision as possible whilst making this mesh with ESS, as he did overall do a good job on these missions and I don't want to just scrap the whole thing. Surprisingly, *most* of this actually meshes just fine with some tweaks. 

This *should* be playable now but frankly I haven't tested it yet.