reorder and labels
Reordered the Wazn, Phurad, and Alheka systems so they make more sense in lore. Wazn is now more securely in the paradise worlds, Alheka's naval garrison is now defending the paradise worlds from Northern Union and Pirate incursions. Pearl is now "closer" to Featherweight, and is a nice little nature reserve (and genetics testing/pharmaceuticals lab planet) under the shield of Betelgeuse shipyards. 

Note on labels:

Since the Galaxy is now in color, the labels blend in too well with them. Also, the labels were not accurately depicting their sectors. The Paradise label took up over half of the north, the Core label extended to Earth, so on, so on. This wasn't bad, so much as crowded and misleading. 

I plan on making new labels that properly show the sector lines, or asking an actual artist to do so.