continuing the story
Phil did most the writing for this part, I added... a lot to it. 

This is actually a deceptively simple mission. It's about as close as Endless Sky gets to a cutscene. You escort a Baccy and two Mules from valhalla to betelgeuse, while you learn a little about the Free Worlds Wolfpack, see that they dont get along with the northern union, then get to see exactly why the Veles class is so important to the oathkeeper vs northern union conflict.  

I did each of the NPC fleet spawns for the sake of visual storytelling. However, due to the RNG aspect of fleet spawns themselves, this may require additional future balancing. Please feel free to submit testing feedback on the difficulty of this specific mission.