Add NPC personality, minor story fixes
One of the things missing from the story in general are opportunities for <first> <last> and the NPCs to be... human. In real life, everyone has a different things that impress them and draw them towards people. For Shah, she will project being cold in order to avoid getting in trouble with her superiors, but it doesn't take much at all to break that shell. 

Added opportunities to break that shell, and bonuses for doing so.