bridge the gap part 1:
RNS intros are now as complete as I, personally, can get them for now. It's up to everyone else to fix and improve this further, unless i somehow get better or need to make changes. 

This should now lead into Phil's Oathkeeper storyline in the north. In theory, this should for all players to test Phil's content without save editing.