10 planet changes by Moonbeam
Per Moonbeam and Phil's discussion, it was decided that the planet formerly known as "Deep" was in need of a rename, do to the fact that there is a faction called The Deep, and references to Deep Space, and a region called The Deep. 

After some conversations with @Janaszar, he suggested the name Roanoke, based on the history of the planet and its similarities to the colony. I obliged.