Reorganize and add governments.
They will now be broken down into sectors, alliances, and groups for ease of working through this

One of the biggest factors breaking ESS is the fact that faction interractions aren't happening as they need to. This is especially true in the far north, and in reference to how the Northern Union responds to Merchants. 

In future updates, i will be heavily abusing the Display Names as things go on.