One addition, three changes
So everyone remember the flame thrower? Because I sure didn't. It was a nice, overpowered, pretty, but entirely non-memorable gun that only the most min-maxy of players (myself included) ever used. And even then we only used it until we were able to overpower everything with raw DPS. 

So, well, I of course got rid of it for ESS planning on fixing it later because i couldnt balance it. 

AND THEN I FORGOT ABOUT IT

But then i was reminded. I looked at this forsaken thing and remember everything that I thought was wrong with it, and set about correcting it

**__And correct it I did__**

This weapon now does a stupid amount of slowing damage, has a huge firing arc, causes some small amount of ion damage, and costs a stupid amount of electricity. 

I call it: The Ion Rain (No, not that Ion Rain shush.)


I also asked a random member of the ESS discord (hydragyrum) to make a unilateral decision for me. I posed an argument on the utility of the beam laser line, and why I was considering moving it to being a secondary weapon. This member agreed with my argument and told me to do it. I am a sucker for peer pressure, and so now its a series of secondary weapons. This series is awaiting actual icons now, but i will be adding a couple of placeholder icons in the mean time