Add Reputation Changes, decrease difficulty.
A common complaint is the difficulty of obtaining Surveillance Pods in the Free Worlds intro. With these changes, you can still obtain the parts normally or, if you know where to look, can rip them out of derelict Thea drones. 

I will also be adding the Pods to pirate sales