OK LOOK SO I KNOW IT SOUNDS CRAZY
But hear me out, by this point in time the Navy/Oathkeepers and the Deep have been independently working with the Free Worlds to clear the Free Worlds of any wrongdoing during the course of the war. In fact, in this mission, the Deep tells the Cap-i-taaaan to to check out the deep sky shit. Well thats navy military equipment. So shouldnt that be locked behind the navy license? (why yes, yes it should be, thank you for catching this.) Since the Free Worlds and the rest of the universe have kinda reconciled, and since first last is besties with a bunch of people who have the authority to give first last a navy loicense, why wouldnt first last get a navy loicense here in rec. 

granted they still wouldnt get it in check

Which means, during the same branch for the RNOathkeepers, wouldnt First Last get access to Free Worlds loicenses, too? Why, yes, yes they would. 

So this is future proofing for the southern republic storyline.