Variants by Hydra
We will be completely revamping variants and fleets, hydragyrum took the first steps on bringing this to fruition. 

I simply reorganized them, and fit them into the old variants.txt so that the game shouldnt break while we wait on more. (Also, chances are we'll be using many of the old variants anyway)