update unlocks
The unlocks now care about what difficulty level you selected, with everyone getting access to smugglers choice, but northern hotrods and southern customs being restricted to easy or hard. 

Made sure to hide the shipyards behind an event that only triggers upon getting the tokens/licenses so as to not confuse new players, or overwhelm the sales board on a player's map.