Remove navy from southern space
MZ communicated that the Free Worlds was splitting because they felt underprotected. The best way I can think of to properly communicate this is to make the player feel like the Navy is just never in the south. 

What better way to do that than to just remove the whole of any navy fleets in the south from the moment the game starts? So that's what I did. The south should now feel properly isolated, justifying the  "Southern Mutual Defense Pact" properly taking over, before becoming the Free Worlds.