Remove redundant fleet
This fleet is going to be completely and only modified by events, so having two versions of the same fleet was redundant. 
This also probably would have caused a lot more work, problems, and headache down the road. But hey, at the very least I had the fleet set up already to work off of it!