Cut Marauders until later date
Maurauders are proving more effort than they are worth in the current stage. The plan is to re-add them at a later date, but to cut them from the core game for now. 

Reasoning:

They are balanced more towards tier 1.5
It will be good to get some quick content out later
They're a headache to modify with everything else we have already been doing