Cut Baccy Story (community suggested)
So after agonizing for days over how to fix this, a handful of ESS community members just decided: Cut it. 

As the Baccy is now actually a really well balanced ship, there's no point in requiring a mission chain to unlock it. Instead, just add the city ship license to the sales at Valhalla for a high cost. Problem solved. 

And so, the problem has indeed been solved, the baccy story has been cut completely, and I fixed everything else so it should work in ESS just fine.