Remove Unique Fleets until Project Alexandria.
Unique fleets are going to be an insane amount of work and detailing, which will require all main plots to be in the game, working, and functional. 

Because of this, it is currently just extra data being taken up for no good reason, and this will be added back in along side marauders, the paragon arc, the ursa polaris revisions, marauders, and the like at a later date.