Add price for city ship license
Since it was suggested/demanded that I just cut the baccy story content, I implemented a pay-gate for the City Ship License as a further deterrent from using the ship. This, when also combined with it's other balancing factors, will keep the ship locked away for only people who *really* want one. 

This also justifies the lack of Baccys and mules in Free Worlds space, and elsewhere.