NUS first half rough draft by Philemon and Kiko
I promise we have actually been working on this game.

Anyway the first draft is buggy beyond playability and unfinished. However, in the interest of keeping a backups, it seems like a good idea to merge this to the workfloor version of the game. 

There are at least 2 more proper missions that need done (by me) as well as error fixes and additional mission slots which are likely to extend this further. 

Phil is moving on to finishing the two RNS branches while I assume full responsibility for the NUS from here in.