Minor formatting updates
I timelined out each of the individual storylines (not applicable on github) and have begun the process of correcting minor issues in format, consistency, and canonical progression, so that it will be easier to finish the storylines. 

The current objective is to get all of the current WIP storylines that already exist up to a playable state, before I write up mission templates and hand them off to @Philemon121 for dialogue writing.