Profit sharing for captured vessels (#170)
### Context

Money is easy to make, and the player doesn't have a lot in the way of
overheads. This makes it very simple for a player to make a lot of money
in a short space of time, which allows them to quickly amass a powerful
fleet and do away with the game's sense of challenge.

We've been adding mechanics to help mitigate this issue. For the last
few versions, commodity trading has required the player to pay profit
shares to the various crew across the fleet. This helps slows down
economic advancement once the player has amassed some power, but leaves
it relatively unhindered in the early game.

We need to extend profit sharing to other kinds of income, such as
capturing and plundering enemy ships. Currently, capturing and
plundering are by far the most powerful way to make money. Capturing is
far more profitable than plundering, especially in the early game.

### Changes

This commit makes capturing and plundering add an outstanding debt to
the player's bank account. The amount of the debt is equal to the value
of the captured or plundered assets, multiplied by the fleet's current
profit share ratio.

### Considerations

As of this commit, the interest rate and term of the profit sharing debt
are hard-coded. In a future commit, we intend to make those values based
on some settings in the data files.

In the future, we want to introduce system that tracks the morale of
each ship in the fleet and creates incentives for the player to keep
their crew alive and happy. As part of that, we intend to punish the
loss of life that often takes place when inherent in capturing ships by
having it create morale problems.

We plan for the profit sharing debts to also affect morale. How it does
that exactly will be up to the data files. At this stage, we think that
having outstanding profit sharing debt would cause morale to tick down
each day, but paying the profit shares up front would trigger a morale
boost. Being unable to pay the daily rates on the profit sharing debt
would of course be a morale disaster.