Add profit sharing for mission payments (#172)
### Context

We have been adding profit sharing to the various ways that the player
can make money. This slows down the player's rate of expansion in the
mid-to-late game and creates incentives to have a smaller fleet.

Smaller fleets make the game's combat more challenging and meaningful.

### Changes

This commit adds profit sharing to mission payments.

### Considerations

The fundamental reason why capturing and plundering are better is that
missions don't pay enough by comparison to the price of a ship. We need
to make missions pay more. Like, a lot more. Adding a zero to the end of
the payment might be about right.