Reduce crew salaries because profit is shared (#173)
### Context

We recently overhauled the crew salaries and fleet shares to make the
game more economically challenging. In doing so, we revealed that
missions don't pay well enough, and that daily salaries are not as fun
as profit shares are. This commit addresses the latter issue.

Overall, we want the player to feel like their fleet costs them
something each day. However, we don't want to to tie the player's hands
and punish them for spending several days doing something fun like
exploring the galaxy.

### Changes

This commit significantly reduces the salaries of all crew members:

> Regular Crew: 100 -> 80
> Pilot: 500 -> 150
> Junior Officer: 1000 -> 200
> Senior Officer: 2000 -> 300
> Executive Officer: 10,000 -> 1,000

### Considerations

Under this model, a person would choose to become a crew member on a
starship not because of the daily pay, which might be better working on
a planet, but for the potential profit shares. They become the focus.

I think that this model lays a better foundation for the upcoming morale
system. If crew members don't get a huge daily salary, then shared
profits are therefore their main motivator. That opens the door for
their morale to be more variable.

When they're making a lot of money because the fleet is smart and
productive, they'll be a lot happier.
  
When the crew aren't seeing enough profits to justify their risks,
they'll become unhappy, be less motivated to run the ship well, and
eventually get mad enough to mutiny.